Completing work started by ghostintranslation
Many changes and additions to bring the initial port up to 95% completion.  Changes sample rate to 44100 from 96k which reduces CPU load by more than 50%. Changed sample block size from 256 to 128 which is the default for the Teensy Audio library and is required for some functions e.g. FFT analysis.